Major change in how components are added to instrument files.
McStasScript now reads component files from the McStas installation
when adding a component, and is thus aware of the allowed parameters,
default values, units and even descriptions harvested from comments.
This is used to create a subclass of component for each component
type added to the instrument.
The user can now access paramters in a component directly as
attributes. The user is informed if parameters are misspelled as
addition of new attributes are locked at the end of __init__.

There are also more help functions focused on helping the user find
the right component and explaining the parameters. These have gotten
some syntax highlighting. The print_long function also reminds the
user of required parameters that have yet to be specified.

The new features are demonstrated in a new jupyter notebook called
McStasScript_demo.ipynb.

The documentation will be updated to reflect the changes very soon.