Restructured the code to a package instead of a module.
The code is now split over multiple files instead of one monolithic
file, making it easier to find to manage the project.

Typical use will now only import a few classes and functions that
are found in the interface package. This includes the instrument
class, plotting and a few utility functions.

The data package contains the types for handling McStas data.

The remaining code is in the helper package.

examples have been moved to an example folder.