Added tests for large parts of the package.
The tests are in the tests package and use unittest.
All tests can be performed using:
  python3 -m unittest discover

The tests covers the data package and most of the helper package.
In the helper package only the formatting and a few methods of
the component class lacks tests.
The interface package is not covered yet.