Minor updates:
Added possibility for incrementing a number after the foldername
instead of failing a run_full_instrument call. This option is
called incremente_folder_name.

Added function to load existing McStas data as McStasData objects,
it is in the functions.py file and called load_data. Test for this
function is included.

Added show_parameters method to instrument so users can see the
available parameters in an instrument model. A test for this
method is included.

A bug was fixed in component_reader where spaces before the
SETTING PARAMETERS keyword made the file reader fail. A strip()
call was added.

Updated the manual to reflect new additions.