Added a configuration file called configuration.yaml.
Currently it holds the paths to the McStas directory, the path to mcrun
and the maximum line length to use when printing to the terminal.

Code has been added to show_parameters (for both instr and comp) that
breaks line dynamically when long comments are printed. In addition
the print_components method prints each component on 1, 2 or 3 lines
depending on the longest lines to maintain a nice overview in most
cases.

Tests were added for the new cases and of reading the configuration
file itself.

The manual was updated with instructions on updating the configuration
file.