Initial work on animation plotting routine.
Update for name_search so that it returns a list if there are multiple matches.
This frequently happens with monitors that produce more than one graph.

The configruation file was moved due to preperations for McStasScript becoming
available on pip. Further work will be needed on configuration when installed
using pip.

Unittests updated to reflect above, and new tests added where necessary.