Major revision, added instrument reading system.
The instrument reader is included in interface as reader. It is
capable of reading McStas instruments and incorparating the
contained information into a McStasScript instance of an
instrument. It is also possible to write a python file to disk
which uses McStasScript to reproduce the given instrument file.

This feature has been tested with 20 McStas instruments from the
example folder in McStas 2.5, where all are reproduced accurately.
The feature should however still be used with some caution as
it is impossible to test for all possible syntax styles and
keyword combinations. It is highly recommended to run a test and
compare the results to the original instrument.

There are some cases where the reader is known to fail:
  Use of c unions (very rare in McStas)
  Use of multidemensional arrays in declare (1 dimension OK)

In the development of this major feature, several smaller features
were required. These are all documented in the manual.
  Possible to copy components on the instrument level
  Possible to add c code before/after components (%include)
  ROTATED statements are now only written if necessary
  Possible to append lines to declare (useful for structs, functions)

All features are tested in unit tests. Integration tests for full
instruments using the reader will follow.

The code is also distributed over pip and all files needed to for
the distribution is included here for completeness.