Fixed a few bugs in read_trace for instrument reader.
First bug produced double ,, when only one was needed.

Second bug added "\" to the instrument file writen, when it was only
supposed to be part of the python file.

Third bug included quotation marks around all group names, which again
was only supposed to be included in the python file.

Test were updated to take this into account.