Start of work on introducing type safety.
Parameters will always have a type, before default was "".
Wonder if I should write double or not in instr files, as it is
default anyway. A lot of unit tests needs to be fixed.

Components have access to instrument parameters and variables.
They check when a component attribute is set if it is a legal
type, and if a parameter / variable is used, it checks the types
matches.

Work in progress.