Added more advanced Variable class which is a super for specialized variables:
	Parameter
	DeclareVariable
	DefinitionParameter
	SettingParameter

The first two are on instrument level, the latter on component level.
They include advanced error checking, including the ability to check
if assigning for example a DeclaredVariable to a SettingParameter
is legal according to the underlying C types.

These classes are not yet used, but will be implemented in the code
base in the coming commits.

This removes a lot of the error checking from the component class.