Further implementation of the new Variable bass class from which the different variables used are derived. We are now using
	Parameter
	DeclareVariable
	DefinitionVariable
	SettingVariable

Which all have their own type checking performed when assigned and
changed. They include information on comments, default value, type
and so on.

The bugs have been ironed out, so now the module works as before
this work started, but with stronger type checking and it is now
easier to update type rules.

All integration tests and unit tests passes.
The test of instrument reader for a number of the McStas examples
passes, but with a few notable issues reported here:

A few issues with the current branch:
	consts are not recoignized
	{} initialization of definition parameters in components
		are counted as an error, but it is in fact allowed
	if using add_declare_line to declare structs or other
		complex variables, the type checking will probably
		report this as an error

Need to add unit tests that specifically test the new features
and saftey checks before merging with master branch.