Allowing saving of created plots instead of displaying them by providing a filename to the plotting functions.
When force_compile is set to False, a new instrument file will
not be written anymore to avoid compile happening because the
new instrument file is newer than the old one. May need to add
a check that ensures the instrument object is unchanged from the
previous, for example by writing a dummy instrument file and
ensuring they differ only in generated time.