Addressing review comments from Celine Durniak, much appreciated!
The review concerns McStas objects and their tests.

Fixed typos in docstrings and comments.

Documented missing methods in docstrings.

Added input sanitation for parameters, declare variable and some
component methods.

Added the option of specifying a float to set_AT instead of a list,
as the most common use is just a distance in the z direction.

Simplified path handling in component_reader

Converted to built-in startswith method instead of adding my own method.

Added a test for checking _freeze and _unfreeze methods on component.

Added test for SPLIT number being a variable.

Declare variable and parameter had different behavior in terms of
spacing after type, this was simplified so only one convention is
used, which incliudes no additional spaces.

Ensured that tests fail when targets are changed after concern
expressed by reviewer.