Update of plotting and simulation GUI along with unit tests.
Now include the live simulation, chopping the simulation into
smaller chunks that are displayed to the user earlier.

Unit tests now test a large part of the GUI code and underlying
methods that are called when using the widgets.

An integration test going through the GUI simulation interface
through the closest method to pressing the run button has also
been added.