Addition of new datatype, event data.
Data loader now checks if Errors are available, if they are a
McStasDataBinned dataset is made, if not it is assumed the data
is events. In that case an McStasDataEvents dataset is created.

Added data set to test_data_set to test loader. Unit tests modified
to take this into account. New test added for Event data.