Fixes to GUI system.
Fixed problem with colorbar axes being visible when they shouldn't.
Fixed problem when switching between runs with and without mpi.
Fixed problem with use of depricated keyword in plotting.

Removed live mode as it was failing.
Still some issues with plotting in simulation widget, since it is
working with the plotting interface it is most likely from threading.

Added ipywidgets to requirements.