Use of libpyvinyl Parameter class and CalculatorParameters class.
Use these as baseclasses for the appropriate classes in McStasScript,
and have made all required changes for the package to be functional.
Unit tests are updated and everything seems to be in order.

Am using direct access to underlying attributes a lot, so it is
hard to change the underlying implementation in libpyvinyl without
breaking this implementation. Will probably need a better contract
with libpyvinyl to be more confident in the features I use will not
be changed.