Code for error checking and remove/move components.
Now components checks parameters that are asigned a string containing
only letters against the instrument parameters and variables, if no
match is found, it raises a McStasError. Probably a misspelling, or a
forgotten set of " / '.

Introduced instrument.has_error() that returns True if an error is
found with instrument.check_for_errors. has_error does not raise
any exceptions.

Added methods for moving and removing components in the instrument.
They both check if the instrument had errors before and after the
operations, if the operation introduced an error it warns the user.

Added a setting called checks that can be set to False if these
kinds of checks should not prevent writing the instrument file.