Adding support for USERVARS which is available in McStas 3.X. Includes showing these in show_variables (needs documentation)
For use in McStas 2.X instruments, USERVARS are not allowed. A
method was added which moves all USERVARS to DECLARE, which for
McStas 2.X would have the same function. Needs documentation.

Introduced checks for avoiding having same name in declare, uservars
and instrument parameters, so each of these are now forced to not
have clashes in namespace, which is important.

Also included new method to display the current instrument code.

Added unit tests for clash detection and USERVARS section.