First version of MCPL bridge.
A subsection of the instrument can be executed.
instr.run_from(component_reference) sets start.
instr.run_to(component_reference) sets end.

MCPL input and output components are auto generated.
It is checked if the run_from / run_to points chosen are allowed,
if there are RELATIVE references across the files it wouldn't work,
so that is simply not allowed and an error is shown that tells
the user what needs to be fixed.

Does not work yet, the mcpl files are written in folders and with
a compression. Work in progress.