Update of diagnostics, common aspects of a diagnostics instrument has been collected in diagnostics_instrument, which will now only have code relevant to keeping a seperate copy of an instrument with common corrections such as removing use of PREVIOUS and updating target_index after insertion of new components.
The event recording system is now called beam_diagnostics and
inherits from diagnostics_instrument.
The intensity analysis is called intensity_diagnostics, and is
now accessed through the instrument diagram that is used to display
the component order. Use show_diagram(analysis=True).

The instrument diagram was updated as well, it now shows connections
for target_index and shares color with JUMP. Legend is kept up to date.