Added search and dependency keywords.
set_dependency on instrument, reset with set_dependency("").

add_search on instrument (handled with a SearchStatement and SearchStatementList class)
add_search on component

In a search statement it is possible to set SHELL to True as keyword argument to get that feature.

Both component and instrument has:
show_search to display current search statements
clear_search to reset the list to empty